requirements.txt
git+https://github.com/hunkim/streamlit-google-oauth
protobuf<=3.20.3 
streamlit
httpx-oauth==0.3.5
python-dotenv
typing_extensions
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client